Change vallidation method for Lenovo BIOS passwords
The operation SetBiosPassword needs a reboot to be efective. After the first execution, all SetBiosPassword operations will fail even if the password is correct until the reboot.
This can create a problem if you do multiple Manage-LenovoBiosSettings operations without rebooting between them.
The method for checking if the password is correct is changed to trying to set a value with it's current value to try to avoid this situation